+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 # [Unreleased]
 
+### Added
+- `--show-only-errors-with-codes X Y Z...` allows specifying a list of error codes, only error messages that matches any of the specified error codes are displayed.
+
 ### Fix
 - The initial TDH in an ITS payload was always interpreted as `no_data = 0`, now also allows `no_data = 1` which allows trigger echoing immediately after an IHW.